+++ autoware_lanelet2_extension/test/src/test_normalize_radian.cpp
@@ -1,0 +1,32 @@
+// Copyright 2025 Autoware Foundation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// NOLINTBEGIN(readability-identifier-naming)
+
+#include "../../lib/normalize_radian.hpp"
+
+#include <gtest/gtest.h>
+
+TEST(normalize_radian, normalize_radian_test)
+{
+  EXPECT_FLOAT_EQ(lanelet::utils::impl::normalize_radian(M_PI * 1.5), M_PI * (-0.5));
+}
+
+int main(int argc, char ** argv)
+{
+  testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
+
+// NOLINTEND(readability-identifier-naming)
